removal of the environment factory concept, simulation start time moved to params, station central now search only for nearest station, not nearest empty, statistic of average km per vehicle fixed,
Former-commit-id: 1d90a42dcc88154b40fed09052481803f8f29656

+++ src/main/java/com/mycompany/testsim/statistics/Statistics.java
@@ -141,10 +141,10 @@ public class Statistics extends EventHandlerAdapter implements SimulationFinishe
     }
 
     private void countAveragesFromAgents() {
-        int metersWithPassengerSum = 0;
-        int metersToStartLocationSum = 0;
-        int metersToStationSum = 0;
-        int metersRebalancingSum = 0;
+        long metersWithPassengerSum = 0;
+        long metersToStartLocationSum = 0;
+        long metersToStationSum = 0;
+        long metersRebalancingSum = 0;
         
         
         for (OnDemandVehicle onDemandVehicle : onDemandVehicleStorage) {
